Fix and adds methods to ytsync-sources
Fix for cycles in all element.
Adds method setPlayerVideId for set the video a id of a player.
Adds method setPlayerVolume for set the volume of a player.
Adds method setPlayerMute for toggle mute of a player.
Adds method addPlayerDelay for add delay to a player.
Adds method setPlayersStart for set the first second of all players.